+++ analysis/import/UnderstandImporter/src/main/kotlin/de/maibornwolff/codecharta/importer/understand/UnderstandImporter.kt
@@ -24,8 +24,8 @@ class UnderstandImporter : Callable<Void> {
     @CommandLine.Option(names = ["-o", "--output-file"], description = ["output File (or empty for stdout)"])
     private var outputFile: File? = null
 
-    @CommandLine.Option(names = ["-c"], description = ["compress output File to gzip format"])
-    private var compress = false
+    @CommandLine.Option(names = ["-nc", "--not-compressed"], description = ["save uncompressed output File"])
+    private var compress = true
 
     @CommandLine.Parameters(arity = "1..*", paramLabel = "FILE", description = ["Understand csv files"])
     private var files: List<File> = mutableListOf()